+++ interpn/benches/bench.rs
@@ -2,7 +2,14 @@
 
 use criterion::*;
 use gridgen::*;
-use interpn::{multicubic, multilinear, MultilinearRegular};
+use interpn::{
+    multicubic, multilinear,
+    one_dim::{
+        hold::{Left1D, Nearest1D},
+        Interp1D,
+    },
+    Linear1D, LinearHoldLast1D, MultilinearRegular, RectilinearGrid1D, RegularGrid1D,
+};
 
 enum Kind {
     Interp,
@@ -207,6 +214,7 @@ fn bench_interp(c: &mut Criterion) {
     for gridsize in [100, 1000] {
         let mut group = c.benchmark_group(format!("Interp_1D_Shuffled_{gridsize}-grid"));
         for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
             bench_interp_specific!(group, 1, gridsize, size, Kind::Interp);
         }
         group.finish();
@@ -215,6 +223,7 @@ fn bench_interp(c: &mut Criterion) {
     for gridsize in [100, 1000] {
         let mut group = c.benchmark_group(format!("Interp_2D_Shuffled_{gridsize}-grid"));
         for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
             bench_interp_specific!(group, 2, gridsize, size, Kind::Interp);
         }
         group.finish();
@@ -223,7 +232,192 @@ fn bench_interp(c: &mut Criterion) {
     for gridsize in [10, 100] {
         let mut group = c.benchmark_group(format!("Interp_3D_Shuffled_{gridsize}-grid"));
         for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
             bench_interp_specific!(group, 3, gridsize, size, Kind::Interp);
+        }
+        group.finish();
+    }
+
+    // 1D specialized linear rectilinear
+    for gridsize in [10, 1000] {
+        let kind = Kind::Interp;
+        let ndims = 1;
+        let mut group = c.benchmark_group(format!("Interp_1D_Special_{gridsize}-grid"));
+        for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
+            group.bench_with_input(
+                BenchmarkId::new(format!("Linear1D Rect. {}-grid", gridsize), size),
+                size,
+                |b, &size| {
+                    // Interpolation grid with noise
+                    let (grids, z) = gen_grid(ndims, gridsize, 1e-3);
+                    let grid = RectilinearGrid1D::new(&grids[0], &z).unwrap();
+
+                    // Observation grid
+                    let m: usize = ((size as f64).powf(1.0 / (ndims as f64)) + 2.0) as usize;
+                    let gridobs_t = match kind {
+                        Kind::Interp => gen_interp_obs_grid(&grids, m, true),
+                        Kind::Extrap => gen_extrap_obs_grid(&grids, m, true),
+                    };
+                    let obs: Vec<&[f64]> = gridobs_t.iter().map(|x| &x[..size]).collect();
+                    let mut out = vec![0.0; size];
+
+                    b.iter(|| {
+                        black_box({
+                            let interp = Linear1D::new(grid);
+                            interp.eval(&obs[0], &mut out).unwrap()
+                        })
+                    });
+                },
+            );
+        }
+        group.finish();
+    }
+
+    // 1D specialized linear regular
+    for gridsize in [10, 1000] {
+        let kind = Kind::Interp;
+        let ndims = 1;
+        let mut group = c.benchmark_group(format!("Interp_1D_Special_{gridsize}-grid"));
+        for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
+            group.bench_with_input(
+                BenchmarkId::new(format!("Linear1D Reg. {}-grid", gridsize), size),
+                size,
+                |b, &size| {
+                    // Interpolation grid with noise
+                    let (grids, z) = gen_grid(ndims, gridsize, 0.0);
+                    let grid =
+                        RegularGrid1D::new(grids[0][0], grids[0][1] - grids[0][0], &z).unwrap();
+
+                    // Observation grid
+                    let m: usize = ((size as f64).powf(1.0 / (ndims as f64)) + 2.0) as usize;
+                    let gridobs_t = match kind {
+                        Kind::Interp => gen_interp_obs_grid(&grids, m, true),
+                        Kind::Extrap => gen_extrap_obs_grid(&grids, m, true),
+                    };
+                    let obs: Vec<&[f64]> = gridobs_t.iter().map(|x| &x[..size]).collect();
+                    let mut out = vec![0.0; size];
+
+                    b.iter(|| {
+                        black_box({
+                            let interp = Linear1D::new(grid);
+                            interp.eval(&obs[0], &mut out).unwrap()
+                        })
+                    });
+                },
+            );
+        }
+        group.finish();
+    }
+
+    // 1D specialized linear hold-last regular
+    for gridsize in [10, 1000] {
+        let kind = Kind::Interp;
+        let ndims = 1;
+        let mut group = c.benchmark_group(format!("Interp_1D_Special_{gridsize}-grid"));
+        for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
+            group.bench_with_input(
+                BenchmarkId::new(format!("LinearHoldLast1D Reg. {}-grid", gridsize), size),
+                size,
+                |b, &size| {
+                    // Interpolation grid with noise
+                    let (grids, z) = gen_grid(ndims, gridsize, 0.0);
+                    let grid =
+                        RegularGrid1D::new(grids[0][0], grids[0][1] - grids[0][0], &z).unwrap();
+
+                    // Observation grid
+                    let m: usize = ((size as f64).powf(1.0 / (ndims as f64)) + 2.0) as usize;
+                    let gridobs_t = match kind {
+                        Kind::Interp => gen_interp_obs_grid(&grids, m, true),
+                        Kind::Extrap => gen_extrap_obs_grid(&grids, m, true),
+                    };
+                    let obs: Vec<&[f64]> = gridobs_t.iter().map(|x| &x[..size]).collect();
+                    let mut out = vec![0.0; size];
+
+                    b.iter(|| {
+                        black_box({
+                            let interp = LinearHoldLast1D::new(grid);
+                            interp.eval(&obs[0], &mut out).unwrap()
+                        })
+                    });
+                },
+            );
+        }
+        group.finish();
+    }
+
+    // 1D specialized hold-left
+    for gridsize in [10, 1000] {
+        let kind = Kind::Interp;
+        let ndims = 1;
+        let mut group = c.benchmark_group(format!("Interp_1D_Special_{gridsize}-grid"));
+        for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
+            group.bench_with_input(
+                BenchmarkId::new(format!("Left1D Reg. {}-grid", gridsize), size),
+                size,
+                |b, &size| {
+                    // Interpolation grid with noise
+                    let (grids, z) = gen_grid(ndims, gridsize, 0.0);
+                    let grid =
+                        RegularGrid1D::new(grids[0][0], grids[0][1] - grids[0][0], &z).unwrap();
+
+                    // Observation grid
+                    let m: usize = ((size as f64).powf(1.0 / (ndims as f64)) + 2.0) as usize;
+                    let gridobs_t = match kind {
+                        Kind::Interp => gen_interp_obs_grid(&grids, m, true),
+                        Kind::Extrap => gen_extrap_obs_grid(&grids, m, true),
+                    };
+                    let obs: Vec<&[f64]> = gridobs_t.iter().map(|x| &x[..size]).collect();
+                    let mut out = vec![0.0; size];
+
+                    b.iter(|| {
+                        black_box({
+                            let interp = Left1D::new(grid);
+                            interp.eval(&obs[0], &mut out).unwrap()
+                        })
+                    });
+                },
+            );
+        }
+        group.finish();
+    }
+
+    // 1D specialized nearest
+    for gridsize in [10, 1000] {
+        let kind = Kind::Interp;
+        let ndims = 1;
+        let mut group = c.benchmark_group(format!("Interp_1D_Special_{gridsize}-grid"));
+        for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
+            group.bench_with_input(
+                BenchmarkId::new(format!("Nearest1D Reg. {}-grid", gridsize), size),
+                size,
+                |b, &size| {
+                    // Interpolation grid with noise
+                    let (grids, z) = gen_grid(ndims, gridsize, 0.0);
+                    let grid =
+                        RegularGrid1D::new(grids[0][0], grids[0][1] - grids[0][0], &z).unwrap();
+
+                    // Observation grid
+                    let m: usize = ((size as f64).powf(1.0 / (ndims as f64)) + 2.0) as usize;
+                    let gridobs_t = match kind {
+                        Kind::Interp => gen_interp_obs_grid(&grids, m, true),
+                        Kind::Extrap => gen_extrap_obs_grid(&grids, m, true),
+                    };
+                    let obs: Vec<&[f64]> = gridobs_t.iter().map(|x| &x[..size]).collect();
+                    let mut out = vec![0.0; size];
+
+                    b.iter(|| {
+                        black_box({
+                            let interp = Nearest1D::new(grid);
+                            interp.eval(&obs[0], &mut out).unwrap()
+                        })
+                    });
+                },
+            );
         }
         group.finish();
     }
@@ -236,6 +430,7 @@ fn bench_extrap(c: &mut Criterion) {
     for gridsize in [10] {
         let mut group = c.benchmark_group(format!("Extrap_1D_Shuffled_{gridsize}-grid"));
         for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
             bench_interp_specific!(group, 1, gridsize, size, Kind::Extrap);
         }
         group.finish();
@@ -244,6 +439,7 @@ fn bench_extrap(c: &mut Criterion) {
     for gridsize in [10] {
         let mut group = c.benchmark_group(format!("Extrap_2D_Shuffled_{gridsize}-grid"));
         for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
             bench_interp_specific!(group, 2, gridsize, size, Kind::Extrap);
         }
         group.finish();
@@ -252,6 +448,7 @@ fn bench_extrap(c: &mut Criterion) {
     for gridsize in [10] {
         let mut group = c.benchmark_group(format!("Extrap_3D_Shuffled_{gridsize}-grid"));
         for size in [1, 100, 1_000_000].iter() {
+            group.throughput(Throughput::Elements(*size as u64));
             bench_interp_specific!(group, 3, gridsize, size, Kind::Extrap);
         }
         group.finish();